fix for legacy weight sensor calibration
