+++ app/Http/Controllers/Api/MeasurementController.php
@@ -15,12 +15,16 @@ use Moment\Moment;
 use League\Fractal;
 use App\Http\Requests\PostSensorRequest;
 
+use App\Http\Controllers\Api\MeasurementLegacyCalculationsTrait;
+
 /**
  * @group Api\MeasurementController
  * Store and retreive sensor data (both LoRa and direct API POSTs)
  */
 class MeasurementController extends Controller
 {
+    use MeasurementLegacyCalculationsTrait;
+
     protected $respose;
     protected $valid_sensors  = [];
     protected $output_sensors = [];
@@ -74,63 +78,6 @@ class MeasurementController extends Controller
                 return $check_sensor;
         }
         return Response::json('No key found for user', 404);
-    }
-
-    protected function convertSensorStringToArray($data_string)
-    {
-        $out = [];
-        $arr = explode("|", $data_string);
-        foreach ($arr as $str) 
-        {
-            $str_arr = explode(":",$str);
-            if (count($str_arr) > 1)
-                $out[$str_arr[0]] = $str_arr[1];
-        }
-        return $out;
-    }
-
-    protected function calculateWeightKg($sensor, $data_array)
-    {
-        $totalWeight    = 0;
-        $use_separate   = false;
-        $separate_names = ['w_fl', 'w_fr', 'w_bl', 'w_br'];
-
-        foreach ($separate_names as $name) 
-        {
-            if (isset($data_array[$name]) && $data_array[$name] > 0)
-                $use_separate = true;
-        }
-        
-        foreach($data_array as $name => $value)
-        {
-            if (($use_separate && in_array($name, $separate_names)) || (!$use_separate && $name == 'w_v')) // 4 names w_fl, w_fr, w_bl, w_br || 1 combined name w_v
-            {
-                // TODO: Make dependent of SensorDefinition
-
-                $name_offset = $this->last_sensor_measurement_time_value($sensor, $name.'_offset');
-                $name_factor = $this->last_sensor_measurement_time_value($sensor, $name.'_kg_per_val');
-                
-                if ($name_factor === 0)
-                {
-                    // exclude from calculation
-                }
-                else if ($name_factor !== null) // factor available
-                {
-                    $offset = isset($name_offset) ? floatval($name_offset) : 0;
-                    $factor = floatval($name_factor);
-                    $weight = $factor * (floatval($value) - $offset);
-                    $totalWeight += $weight;
-                    
-                    //die("offset=$offset factor=$factor weight=$weight totalWeight=$totalWeight");
-                }
-                else
-                {
-                    $totalWeight += floatval($value);
-                }
-            }
-        }
-        //die(print_r($log));
-        return $totalWeight;
     }
 
     
@@ -212,106 +159,6 @@ class MeasurementController extends Controller
         return $out_arr; 
     }
 
-    private function floatify_sensor_val($arr, $key)
-    {
-        if (isset($arr[$key]))
-        {
-            $value = $arr[$key];
-
-            if ($value == 0)
-            {
-                unset($arr[$key]);
-            }
-            else
-            {
-                switch($key)
-                {
-                    case 't':
-                    case 't_i':
-                        $arr[$key] = ($value / 5) - 10; // de-tempfy
-                        break;
-                    case 'h':
-                        $arr[$key] = $value / 2;
-                        break;
-                    case 'bv':
-                        $arr[$key] = $value / 10;
-                        break;
-                    case 'w_v':
-                        $arr[$key] = $value;
-                        break;
-                    case 'w_fl':
-                    case 'w_fr':
-                    case 'w_bl':
-                    case 'w_br':
-                        $arr[$key] = $value / 300; // de-weightify
-                        break;
-                }
-            }
-        }
-
-        return $arr;
-    }
-
-    private function decode_simpoint_payload($payload)
-    {
-        $out = [];
-        $beep_sensors = [
-            't'  , // 0
-            'h'  , // 1
-            'w_v',
-            't_i',
-            'a_i',
-            'bv' ,
-            's_tot',
-            's_fan_4',
-            's_fan_6',
-            's_fan_9',
-            's_fly_a',
-            'w_fl_hb',
-            'w_fl_lb',
-            'w_fr_hb',
-            'w_fr_lb',
-            'w_bl_hb',
-            'w_bl_lb',
-            'w_br_hb',       
-            'w_br_lb', // 18  
-        ];
-
-        $minLength = min(strlen($payload)/2, count($beep_sensors));
-
-        for ($i=0; $i < $minLength; $i++) 
-        { 
-            if (strlen($payload) > count($beep_sensors)*2)
-            {
-                $index = $i * 4 + 2; 
-            }
-            else
-            {
-                $index = $i * 2;
-            }
-            $sensor = $beep_sensors[$i];
-            $hexval = substr($payload, $index, 2);
-
-            if (strpos($sensor, '_hb') !== false) // step 1 of 2 byte value
-            {
-                $sensor = substr($sensor, 0, strpos($sensor, '_hb'));
-                $out[$sensor] = $hexval;
-            } 
-            else if (strpos($sensor, '_lb') !== false) // step 2 of 2 byte value
-            {
-                $sensor = substr($sensor, 0, strpos($sensor, '_lb'));
-                $totalHexVal  = $out[$sensor].$hexval;
-                $out[$sensor] = hexdec($totalHexVal);
-            }
-            else
-            {
-                $out[$sensor] = hexdec($hexval);
-            }
-        }
-        //die(print_r($minLength));
-        return $out;
-    }
-
 
     // requires at least ['name'=>value] to be set
     private function storeInfluxData($data_array, $dev_eui, $unix_timestamp)
@@ -356,6 +203,29 @@ class MeasurementController extends Controller
         return $stored;
     }
 
+    private function createOrUpdateDefinition($device, $abbr_in, $abbr_out, $offset=null, $multiplier=null)
+    {
+        $measurement_in  = Measurement::where('abbreviation',$abbr_in)->first();
+        $measurement_out = Measurement::where('abbreviation',$abbr_out)->first();
+        if ($measurement_in && $measurement_out)
+        {
+            $def = $device->sensorDefinitions()->where('input_measurement_id', $measurement_in->id)->where('output_measurement_id', $measurement_out->id)->first();
+            if ($def && (isset($offset) || isset($multiplier)) ) 
+            {
+                if (isset($offset))
+                    $def->offset = $offset;
+
+                if (isset($multiplier))
+                    $def->multiplier = $multiplier;
+
+                $def->save();
+            }
+            else
+            {
+                $device->sensorDefinitions()->create(['input_measurement_id'=>$measurement_in->id, 'output_measurement_id'=>$measurement_out->id, 'offset'=>$offset, 'multiplier'=>$multiplier]);
+            }
+        }
+    }
 
     private function storeMeasurements($data_array)
     {
@@ -396,45 +266,35 @@ class MeasurementController extends Controller
                         $data_array[$abbr_o] = $def->calibrated_measurement_value($val);
                     }
                 }
-                else // make new calibration values based on stored ones
+                else if ($abbr == 'w_v') // make new calibration values based on stored ones
                 {
-                    if ($abbr == 'w_v' || $abbr == 'w_fl' || $abbr == 'w_fr' || $abbr == 'w_bl' || $abbr == 'w_br')
-                    {
-                        $influx_offset = floatval($this->last_sensor_measurement_time_value($device, $abbr.'_offset'));
-                        $influx_multi  = floatval($this->last_sensor_measurement_time_value($device, $abbr.'_kg_per_val'));
+                    $influx_offset = floatval($this->last_sensor_measurement_time_value($device, $abbr.'_offset'));
+                    $influx_multi  = floatval($this->last_sensor_measurement_time_value($device, $abbr.'_kg_per_val'));
 
-                        if ($influx_offset != 0 || $influx_multi != 0)
-                        {
-                            $measurement_out = Measurement::where('abbreviation','weight_kg')->first();
-                            if ($measurement_out)
-                            {
-                                $device->sensorDefinitions()->create(['input_measurement_id'=>$measurement->id, 'output_measurement_id'=>$measurement_out->id, 'offset'=>$influx_offset, 'multiplier'=>$influx_multi]);
-                            }
-                        }
-                    }
+                    if ($influx_offset != 0 || $influx_multi != 0)
+                        $this->createOrUpdateDefinition($device, $abbr, 'weight_kg', $influx_offset, $influx_multi);
                 }
             }
             
         }
-        //die(print_r($data_array));
-        
-        // Old offset
-        // if (isset($data_array['w_v']) || isset($data_array['w_fl']) || isset($data_array['w_fr']) || isset($data_array['w_bl']) || isset($data_array['w_br'])) 
-        // {
-        //     // check if calibration is required
-        //     $calibrate = $this->last_sensor_measurement_time_value($device, 'calibrating_weight');
-        //     if (floatval($calibrate) > 0)
-        //         $cal = $this->calibrate_weight_sensors($device, $calibrate, false, $data_array);
 
-        //     // take into account offset and multi
-        //     $weight_kg = $this->calculateWeightKg($device, $data_array);
-        //     $data_array['weight_kg'] = $weight_kg;
+        // Legacy weight calculation from 2-4 load cells
+        if (isset($data_array['w_fl']) || isset($data_array['w_fr']) || isset($data_array['w_bl']) || isset($data_array['w_br']) || isset($data_array['w_v'])) 
+        {
+            // check if calibration is required
+            $calibrate = $this->last_sensor_measurement_time_value($device, 'calibrating_weight');
+            if (floatval($calibrate) > 0)
+                $this->calibrate_weight_sensors($device, $calibrate, false, $data_array);
 
-        //     // check if we need to compensate weight for temp
-        //     $data_array = $this->add_weight_kg_corrected_with_temperature($device, $data_array);
-        // }
+            // take into account offset and multi
+            $weight_kg = $this->calculateWeightKg($device, $data_array);
+            $data_array['weight_kg'] = $weight_kg;
+
+            // check if we need to compensate weight for temp
+            $data_array = $this->add_weight_kg_corrected_with_temperature($device, $data_array);
+        }
         
-        //die(print_r($dev_eui));
+        
         $time = time();
         if (isset($data_array['time']))
             $time = intVal($data_array['time']);
@@ -449,214 +309,6 @@ class MeasurementController extends Controller
             Storage::disk('local')->put('sensors/sensor_write_error.log', json_encode($data_array));
             return Response::json('sensor-write-error', 500);
         }
-    }
-
-
-    private function add_weight_kg_corrected_with_temperature($sensor, $data_array)
-    {
-        if (isset($data_array['t']))
-        {
-            $sensor_names    = ['w_fl', 'w_fr', 'w_bl', 'w_br'];
-            $temp_weight_arr = [];
-            foreach ($sensor_names as $key) 
-            {
-                if(isset($data_array[$key]))
-                {
-                    $temp_weight_arr[$key] = $data_array[$key];
-                }
-            }
-            
-            if (count($temp_weight_arr) == 0) // no value available, so do not add
-                return $data_array; 
-
-            $diff_arr = $this->last_sensor_increment_values($sensor, $data_array);
-
-            if (!isset($diff_arr['t']) || $data_array['weight_kg'] == 0)
-                return $data_array;
-
-            $diff_temp = $diff_arr['t'];
-            
-            // temp -> raw weight value correlation is 2 -> 1 = 0.5 W/T
-            // So apply '- diff_arr['t'] * 0.5' to raw values and then calculate the total weight again (incl. offset and factor)
-            $weight_per_temp = 0.35;
-            
-            // foreach ($temp_weight_arr as $sens => $val) 
-            // {
-            //     if (isset($diff_arr[$sens]))
-            //     {
-            //         // check if the value is dependent in the same direction, otherwise do not correct
-            //         $sign                   = (($diff_temp >= 0 && $diff_arr[$sens] >= 0) || ($diff_temp < 0 && $diff_arr[$sens] < 0)) ? -1 : 0; 
-            //         $temp_weight_arr[$sens] = $val + ($sign * $weight_per_temp * $diff_temp);
-            //     }
-            // }
-            // // calculate corrected total weight
-            // $temp_corr_weight = $this->calculateWeightKg($sensor, $temp_weight_arr);;
-            //die(print_r(['tw'=>$temp_weight_arr, 'diff'=>$diff_arr, 'corr'=>$temp_corr_weight,'uncorr'=>$data_array['weight_kg']]));
-
-            $sign             = (($diff_temp >= 0 && $diff_arr['weight_kg'] >= 0) || ($diff_temp < 0 && $diff_arr['weight_kg'] < 0)) ? -1 : 0; 
-            $temp_corr_weight = $data_array['weight_kg'] + ($sign * $weight_per_temp * $diff_temp);
-
-            if ($temp_corr_weight > 0)
-                $data_array['weight_kg_corrected'] = $temp_corr_weight;
-        }
-        return $data_array;
-    }
-
-
-    private function calibrate_weight_sensors($sensor, $weight_kg, $next_measurement=true, $data_array=null)
-    {
-        if ($next_measurement)
-        {
-            $store  = ['calibrating_weight'=>$weight_kg];
-            //die(print_r($store));
-            $stored = $this->storeInfluxData($store, $sensor->key, time());
-            if ($stored)
-                return true;
-        }
-        else
-        {
-            if ($weight_kg == 0)
-                return 'no-calibration-weight-error';
-
-            $search_for    = ['w_fl'=>0, 'w_fr'=>0, 'w_bl'=>0, 'w_br'=>0];
-            $calibrate     = [];
-            $value_counter = 0;
-
-            // set and count values to use
-            foreach ($search_for as $key => $value) 
-            {
-                $last_value = 0;
-                if (isset($data_array[$key]))
-                    $last_value = $data_array[$key];
-                else
-                    $last_value = $this->last_sensor_measurement_time_value($sensor, $key);
-                
-                $key_offset = floatval($this->last_sensor_measurement_time_value($sensor, $key.'_offset'));
-                $last_value = $last_value - $key_offset;
-
-                if (isset($last_value) && $last_value > 1)
-                {
-                    $calibrate[$key.'_kg_per_val'] = $last_value;
-                    $search_for[$key]              = $last_value;
-                    $value_counter++;
-                }
-                else
-                {
-                    $calibrate[$key.'_kg_per_val'] = 0;
-                }
-            }
-
-            if (count($calibrate) == 0)
-                return 'no-weight-values-error';
-
-            $weight_part = $weight_kg / max(1, $value_counter);
-
-            foreach ($calibrate as $key => $value) 
-                $calibrate[$key] = $value != 0 ? $weight_part / $value : 0;
-
-            // check if w_v should be added
-            $key = 'w_v';
-            $combined_value = 0;
-            if (isset($data_array[$key]))
-                $combined_value = $data_array[$key];
-            else
-                $combined_value = $this->last_sensor_measurement_time_value($sensor, $key);
-
-            if ($combined_value > 0)
-            {
-                $calibrate[$key.'_kg_per_val'] = $weight_kg / $combined_value;
-                $search_for[$key]              = $combined_value;
-            }
-
-            // switch 'off' calibrating mode
-            $calibrate['calibrating_weight'] = 0;
-
-            //die(print_r(['weight_kg'=>$weight_kg, 'values'=>$search_for, 'cal'=>$calibrate,'key'=>$sensor->key]));
-
-            $stored = $this->storeInfluxData($calibrate, $sensor->key, time());
-            if ($stored)
-                return true;
-        }
-
-        return 'calibrate-error';
-    }
-
-    private function offset_weight_sensors($sensor)
-    {
-        $search_for    = ['w_fl', 'w_fr', 'w_bl', 'w_br', 'w_v'];
-        $offset        = [];
-        
-        foreach ($search_for as $key) 
-        {
-            $value = $this->last_sensor_measurement_time_value($sensor, $key);
-            if ($value != null)
-                $offset[$key.'_offset'] = $value;
-        }
-
-        if (count($offset) == 0)
-            return 'no-weight-values-error';
-
-        //die(print_r(['cal'=>$offset,'key'=>$sensor->key]));
-
-        $stored = $this->storeInfluxData($offset, $sensor->key, time());
-        if ($stored)
-            return true;
-
-        return 'offset-error';
-    }
-
-
-    // Public functions
-
-    /**
-    * api/sensors/calibrateweight
-    * Legacy method, used by legacy webapp to store weight calibration value e.g.[w_v_kg_per_val] in Influx database, to lookup and calculate [weight_kg] at incoming measurement value storage
-    *
-    * At the next measurement coming in, calibrate each weight sensor with it's part of a given weight.
-    * Because the measurements can come in only each hour/ 3hrs, set a value to trigger the calcuylation on next measurement
-    *
-    * 1. If $next_measurement == true: save 'calibrating' = true in Influx with the sensor key
-    * 2. If $next_measurement == false: save 'calibrating' = false in Influx with the sensor key and...
-    * 3.   Get the last measured weight values for this sensor key, 
-    *      Divide the given weight (in kg) with the amount of sensor values > 1.0 (assuming the weight is evenly distributed)
-    *      Calculate the multiplier per sensor by dividing the multiplier = weight_part / (value - offset)
-    *      Save the multiplier as $sensor_name.'_kg_per_val' in Influx
-    */
-    public function calibrateweight(Request $request)
-    {
-        $validator = Validator::make($request->all(), [
-            'next_measurement' => 'nullable|boolean',
-            'weight_kg'        => 'numeric|required_if:next_measurement,true',
-        ]);
-
-        if ($validator->fails())
-            return Response::json('validation-error', 500);
-
-        $sensor           = $this->get_user_sensor($request, true); // requires id, key, hive_id, or nothing (if only one sensor) to be set
-        $next_measurement = $request->filled('next_measurement') ? $request->input('next_measurement') : true;
-        $weight_kg        = floatval($request->filled('weight_kg') ? $request->input('weight_kg') : $this->last_sensor_measurement_time_value($sensor, 'calibrating_weight'));
-        $calibrated       = $this->calibrate_weight_sensors($sensor, $weight_kg, $next_measurement);
-
-        if($calibrated === true)
-        {
-            if ($next_measurement)
-                return Response::json("calibrating_weight", 200);
-            else
-                return Response::json("calibrated_weight", 201);
-        }
-        
-        return Response::json($calibrated, 500);
-    }
-
-    public function offsetweight(Request $request)
-    {
-        $sensor = $this->get_user_sensor($request, true); // requires id, key, hive_id, or nothing (if only one sensor) to be set
-        $offset = $this->offset_weight_sensors($sensor);
-
-        if($offset === true)
-            return Response::json("offset_weight", 201);
-        
-        return Response::json($offset, 500);
     }
 
     /**
@@ -680,29 +332,6 @@ class MeasurementController extends Controller
         return Response::json('error', 404);
     }
 
-    /**
-    api/sensors/lastweight GET
-    Request last weight related measurement values from a sensor (Device), used by legacy webapp to show calibration data: ['w_fl', 'w_fr', 'w_bl', 'w_br', 'w_v', 'weight_kg', 'weight_kg_corrected', 'calibrating_weight', 'w_v_offset', 'w_v_kg_per_val', 'w_fl_offset', 'w_fr_offset', 'w_bl_offset', 'w_br_offset']
-    @authenticated
-    @bodyParam key string DEV EUI to look up the sensor (Device)
-    @bodyParam id integer ID to look up the sensor (Device)
-    @bodyParam hive_id integer Hive ID to look up the sensor (Device)
-    */
-    public function lastweight(Request $request)
-    {
-        $weight = ['w_fl', 'w_fr', 'w_bl', 'w_br', 'w_v', 'weight_kg', 'weight_kg_corrected', 'calibrating_weight', 'w_v_offset', 'w_v_kg_per_val', 'w_fl_offset', 'w_fr_offset', 'w_bl_offset', 'w_br_offset'];
-        $sensor = $this->get_user_sensor($request);
-        $output = $this->last_sensor_values_array($sensor, implode('","',$weight));
-
-        if ($output === false)
-            return Response::json('sensor-get-error', 500);
-        else if ($output !== null)
-            return Response::json($output);
-    
-        return Response::json('error', 404);
-    }
-
-    
 
     private function parse_kpn_payload($request_data)
     {
